Fix real .app packaging, permission helper leak, and Preferences window
Adds an export-app.sh script that wraps the SPM executable in a proper
.app bundle (with a real Info.plist and ad-hoc/Developer ID signing),
which Accessibility permission grants and Launch at Login both need
to behave reliably. Fixes a retain-cycle-prone timer capture in
PermissionsHelper and a UInt/UInt64 mismatch in KeyboardManager's
modifier matching.

Also fixes "Preferences…" doing nothing: SwiftUI's Settings scene
shows itself via a responder-chain action that isn't reliably
delivered in a menu-bar-only (.accessory) app with no other window
open. AppDelegate now manages its own NSWindow for Preferences instead.

+++ Sources/InputCustomizer/Managers/KeyboardManager.swift
@@ -56,7 +56,7 @@ final class KeyboardManager {
         for rule in settingsStore.rules(for: .keyboard) {
             guard case let .keyCombo(ruleKeyCode, ruleModifiers) = rule.trigger,
                   ruleKeyCode == keyCode,
-                  UInt(ruleModifiers) == UInt(modifiers) & relevantModifierMask else { continue }
+                  UInt64(ruleModifiers) == modifiers & relevantModifierMask else { continue }
 
             apply(action: rule.action, to: event)
             return Unmanaged.passRetained(event)
